Spring-boot-Assignment

+++ Spring-boot/h2-demo-main/src/main/java/com/javabytes/springh2demo/Repository/EmployeeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.javabytes.springh2demo.Model.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long>{
+public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
 
 }
